requirements.txt
# Dependencias para el Sistema RAG
# Instalación: pip install -r requirements.txt

# Core AI/ML libraries
openai>=1.100.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0

# Vector database
chromadb>=1.1.0

# Document processing
pypdf>=6.0.0
python-dotenv>=1.1.0

# Utilities
tqdm>=4.67.0
numpy>=1.24.0
pandas>=2.0.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
jupyter>=1.0.0
notebook>=6.5.0